@@ -14,20 +14,21 @@
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^1.9.4",
+		"@sveltejs/adapter-static": "^3.0.8",
+		"@sveltejs/kit": "^2.20.7",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@testing-library/svelte": "^5.2.7",
-		"svelte": "^4.2.19",
-		"svelte-check": "^4.1.6",
+		"svelte": "^5.27.0",
+		"svelte-check": "^4.1.5",
 		"svelte-preprocess": "^6.0.3",
-    "tailwindcss": "^3.4.17",
-		"typescript": "^5.8.3",
-		"vite": "^5.4.14",
-		"vitest": "^3.1.1"
+		"tailwindcss": "^3.4.17",
+		"typescript": "^5.8.2",
+		"vite": "^6.2.6",
+		"vitest": "^3.0.9"
 	},
 	"dependencies": {
-		"@sveltejs/adapter-static": "^3.0.8",
-		"@sveltejs/kit": "^2.5.7",
 		"dayjs": "^1.11.13",
-		"firebase": "^11.6.0"
+		"firebase": "^11.5.0"
 	},
 	"engines": {
 		"node": ">=20 <21"
