@@ -5,7 +5,7 @@ import BackToHome from "$lib/components/BackToHome.svelte";
 <div class="mx-auto px-4 md:px-2 py-4 max-w-[800px]">
   <h1 class="font-bold text-red-500 leading-normal text-4xl">べからず集</h1>
   <div class="m-auto px-2 py-1 bg-[#333] max-w-[310px]">
-    <pre class="my-3 text-white leading-tight text-[0.75rem]">
+    <pre class="my-3 text-white leading-tight text-[0.75rem] font-mono">
 ＿人人人人人人人人人人人人人人人人人人人人＿
 ＞  うそはうそであると見抜ける人でないと  ＜
 ＞  （掲示板を使うのは）難しい　　　　　  ＜
